Change usize to u32 where appropriate
Potentially could be reduced even further to u16 in a lot of cases
but u32 is safe while still saving a lot of space. In cases where
the numbers are used for indexing and such I left in usize (even
though generally u32 would also be fine) for convenience purposes.

+++ src/aco/acs.rs
@@ -79,8 +79,8 @@ impl ACSColony<'_> {
     }
 }
 
-fn calculate_initial_values(nn_tour_length: usize, num_nodes: usize) -> f64 {
-    1.0 / (num_nodes * nn_tour_length) as f64
+fn calculate_initial_values(nn_tour_length: u32, num_nodes: usize) -> f64 {
+    1.0 / (num_nodes * nn_tour_length as usize) as f64
 }
 
 fn global_update_pheromones(pheromones: &mut FloatMatrix,
